Windows/MSVC portability fixes
Fixes for cross-platform bugs exposed by the MSVC build, mostly from
type-width differences between LP64 (Linux) and LLP64 (Windows):

- Socket handle truncation: store the handle in a portable socket_t
  (uintptr_t on Windows to match SOCKET, int on POSIX) with an
  INVALID_SOCK sentinel, instead of an int that dropped the upper
  32 bits of a Windows SOCKET (MSVC C4244).
- Segment calculation: divide by (1LL << 31), not (1L << 31). long is
  32-bit on Windows, so 1L << 31 overflowed into the sign bit and
  negated the divisor, producing wrong/zeroed segments and owners.
- Build the Docker/POSIX-only integration tests on non-Windows platforms
  only (they use popen/pclose and bash scripts unavailable on MSVC).

+++ src/hash/ConsistentHash.cpp
@@ -72,7 +72,10 @@ int ConsistentHash::getSegment(const ByteArray& key) const {
     //
     // This is equivalent to: segment = (normalizedHash * numSegments) / 2^31
     // Using 64-bit to avoid overflow
-    int segment = (static_cast<int64_t>(normalizedHash) * numSegments_) / (1L << 31);
+    // NOTE: use 1LL (guaranteed >= 64-bit) not 1L: on Windows (LLP64) long is
+    // 32-bit, so 1L << 31 overflows into the sign bit and the divisor becomes
+    // negative, producing negated/zero segments.
+    int segment = static_cast<int>((static_cast<int64_t>(normalizedHash) * numSegments_) / (1LL << 31));
 
     // Extra safety: ensure segment is valid
     if (segment < 0 || segment >= static_cast<int>(segmentOwners_.size())) {
